added 2 models and fix migrations

+++ backend/migrations/0006_create_order_table.up.sql
@@ -2,7 +2,7 @@ CREATE TABLE orders (
     id SERIAL PRIMARY KEY,
     user_id INT NOT NULL,
     status VARCHAR(50) DEFAULT 'pending',
-    total NUMERIC(10,2) DEFAULT 0,
+    total INT DEFAULT 0,
     delivery_name VARCHAR(255),      -- имя получателя
     delivery_phone VARCHAR(50),      -- телефон получателя
     delivery_address TEXT,           -- адрес доставки
